requirements.txt
streamlit==1.30.0
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
langchain-openai==0.0.2.post1
python-dotenv==1.0.0
weaviate==3.26.1
python-dotenv==1.0.1
pypdf==4.0.1